refactor(test-suite): split the operator edge case tests by group
